Move slim transform helpers into common file
